requirements.txt
python-dotenv==0.19.0
requests==2.26.0
deepl-api==0.2.2